mk1/backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
pyyaml>=5.4.1
opencv-python>=4.5.3
torch>=1.9.0
ultralytics>=8.0.0
transformers>=4.11.0
Pillow>=8.3.1
numpy>=1.21.0
deep-sort-realtime>=1.3.2
scikit-image>=0.18.0
tqdm>=4.62.0 